+++ src/plugins/PatchMatch.js
@@ -10,7 +10,9 @@
 var stdMath = Math, INF = Infinity,
     copy = FILTER.Util.Array.copy,
     clamp = FILTER.Util.Math.clamp,
-    TypedArray = FILTER.Util.Array.typed, TypedObj = FILTER.Util.Array.typed_obj;
+    toJSON = JSON.stringify,
+    fromJSON = JSON.parse,
+    TypedObj = FILTER.Util.Array.typed_obj;
 
 // PatchMatch algorithm filter
 // https://en.wikipedia.org/wiki/PatchMatch
@@ -19,7 +21,7 @@ FILTER.Create({
     name: "PatchMatchFilter"
 
     ,_update: true
-    ,hasMeta: false
+    ,hasMeta: true
     ,hasInputs: true
 
     // parameters
@@ -28,11 +30,14 @@ FILTER.Create({
     // The patch size (should be odd)
     ,patch: 11
     // The radius of the area for the random search
-    ,radius: 200
+    ,radius: 100
     // The random search alpha
     ,alpha: 0.5
-    // patch areas/selections
-    ,areas: null
+    // patch area/selection
+    ,fromArea: null
+    ,toArea: null
+    ,pyramid: false
+    ,average: false
     ,returnNNF: false
 
     ,init: function() {
@@ -42,6 +47,13 @@ FILTER.Create({
     // support worker serialize/unserialize interface
     ,path: FILTER.Path
 
+    ,dispose: function() {
+        var self = this;
+        self.fromArea = self.toArea = null;
+        self.$super('dispose');
+        return self;
+    }
+
     ,params: function(params) {
         var self = this;
         if (params)
@@ -50,7 +62,10 @@ FILTER.Create({
             if (null != params.patch) self.patch = +params.patch;
             if (null != params.radius) self.radius = +params.radius;
             if (null != params.alpha) self.alpha = +params.alpha;
-            if (null != params.areas) self.areas = params.areas;
+            if (null != params.pyramid) self.pyramid = params.pyramid;
+            if (null != params.average) self.average = params.average;
+            if (null != params.fromArea) self.fromArea = params.fromArea;
+            if (null != params.toArea) self.toArea = params.toArea;
             if (null != params.returnNNF) self.returnNNF = params.returnNNF;
         }
         return self;
@@ -63,7 +78,10 @@ FILTER.Create({
         patch: self.patch,
         radius: self.radius,
         alpha: self.alpha,
-        areas: JSON.stringify(self.areas),
+        pyramid: self.pyramid,
+        average: self.average,
+        fromArea: toJSON(self.fromArea),
+        toArea: toJSON(self.toArea),
         returnNNF: self.returnNNF
         };
     }
@@ -74,74 +92,143 @@ FILTER.Create({
         self.patch = params.patch;
         self.radius = params.radius;
         self.alpha = params.alpha;
-        self.areas = JSON.parse(params.areas);
+        self.pyramid = params.pyramid;
+        self.average = params.average;
+        self.fromArea = fromJSON(params.fromArea);
+        self.toArea = fromJSON(params.toArea);
         self.returnNNF = params.returnNNF;
         return self;
     }
 
     ,metaData: function(serialisation) {
+        var self = this;
         if (serialisation && FILTER.isWorker)
         {
-            return TypedObj(this.meta && this.meta.nnf ? {nnf:this.meta.nnf.map(function(nnf) {
-                return patchmatch.NNF.serialize(nnf);
-            })} : this.meta);
+            return TypedObj(self.meta && self.meta.nnf ? {nnf:patchmatch.NNF.serialize(self.meta.nnf)} : self.meta);
         }
         else
         {
-            return this.meta;
+            return self.meta;
         }
     }
 
     ,setMetaData: function(meta, serialisation) {
+        var self = this;
         if (serialisation && ("string" === typeof meta))
         {
-            this.meta = TypedObj(meta, 1);
-            if (this.meta.nnf) this.meta.nnf = this.meta.nnf.map(function(nnf) {
-                return patchmatch.NNF.unserialize(null, null, nnf);
-            });
+            self.meta = TypedObj(meta, 1);
+            if (self.meta.nnf) self.meta.nnf = patchmatch.NNF.unserialize(null, null, self.meta.nnf);
         }
         else
         {
-            this.meta = meta;
+            self.meta = meta;
         }
-        return this;
+        return self;
     }
 
     ,apply: function(im, w, h) {
-        var self = this, areas = self.areas,
-            Area = FILTER.Util.Image.Selection;
+        var self = this,
+            fromArea = self.fromArea,
+            toArea = self.toArea,
+            input, metrics, dst, src, nnf, nnf2,
+            im2, w2, h2,
+            Area = FILTER.Util.Image.Selection,
+            Pyramid = FILTER.Util.Image.Pyramid;
+        self._update = false;
+        self.hasMeta = true;
         if (self.returnNNF)
         {
-            self._update = false;
-            self.hasMeta = true;
-            self.meta = {nnf:[]};
+            self.meta = {nnf:null};
         }
         else
         {
-            self._update = true;
-            self.hasMeta = true;
-            self.meta = {metrics:[]};
+            self.meta = {metric:0};
         }
-        if (areas && areas.length) areas.forEach(function(selection) {
-            if (!selection['from'] || !selection['to']) return;
-            var srcInput = self.input(selection['from'].data), dst, src, nnf, metrics;
-            if (!srcInput) return;
-            src = new Area(srcInput[0], srcInput[1], srcInput[2], 4, selection['from']);
-            dst = new Area(im, w, h, 4, selection['to']);
-            nnf = patchmatch(dst, src, self.patch, self.iters, self.alpha, self.radius);
-            if (self.returnNNF)
+        if (fromArea && toArea)
+        {
+            input = fromArea.data ? self.input(fromArea.data) : [im, w, h];
+            if (input)
             {
-                self.meta.nnf.push(nnf);
+                im2 = input[0]; w2 = input[1]; h2 = input[2];
+                if (im2 === im) im2 = copy(im2);
+                if (self.pyramid)
+                {
+                    dst = (new Pyramid()).build(im, w, h, 4, self.patch, new Area(im, w, h, 4, toArea));
+                    src = (new Pyramid()).build(im2, w2, h2, 4, self.patch, new Area(im2, w2, h2, 4, fromArea));
+                    for (var i=dst.levels.length-1; i>=0; --i)
+                    {
+                        nnf2 = nnf ? patchmatch(
+                            nnf.scale(dst.levels[i].area, src.levels[i].area, 2),
+                            null,
+                            self.patch,
+                            self.iters,
+                            self.alpha,
+                            self.radius
+                        ) : patchmatch(
+                            dst.levels[i].area,
+                            src.levels[i].area,
+                            self.patch,
+                            self.iters,
+                            self.alpha,
+                            self.radius
+                        );
+                        if (nnf) nnf.dispose(true);
+                        nnf = nnf2;
+                    }
+                    dst.dispose(); src.dispose();
+                }
+                else
+                {
+                    nnf = patchmatch(
+                        new Area(im, w, h, 4, toArea),
+                        new Area(im2, w2, h2, 4, fromArea),
+                        self.patch,
+                        self.iters,
+                        self.alpha,
+                        self.radius
+                    );
+                }
+                if (self.returnNNF)
+                {
+                    self.meta.nnf = nnf;
+                }
+                else
+                {
+                    metrics = {error:0, mode:self.average ? "average" : "default"};
+                    nnf.apply(metrics).dispose(true);
+                    self.meta.metric = metrics.error;
+                    self._update = true;
+                }
             }
-            else
-            {
-                nnf.apply(metrics={error:0,mode:selection.mode||"default"}).dispose(true);
-                self.meta.metrics.push(metrics.error);
-            }
-        });
+        }
         return im;
     }
 });
+
+function patchmatch(dst, src, patch, iters, alpha, radius)
+{
+    if (dst instanceof patchmatch.NNF)
+    {
+        return dst.run(iters, alpha, radius);
+    }
+    else
+    {
+        var srcData = src.data(), srcRect = src.rect();
+        if (dst.data().data === srcData.data)
+        {
+            // dst === src, make copy to avoid distortions
+            src = new FILTER.Util.Image.Selection(
+                copy(srcData.data), srcData.width, srcData.height, srcData.channels,
+                {
+                x:srcRect.from.x, y:srcRect.from.y, width:srcRect.width, height:srcRect.height,
+                points:src.points()
+                }
+            );
+        }
+        return (new patchmatch.NNF(dst, src, patch)).init().run(iters, alpha, radius);
+    }
+}
+FILTER.Util.Filter.patchmatch = patchmatch;
 
 function NNF(dst, src, patch)
 {
@@ -173,6 +260,8 @@ NNF.prototype = {
     dstImg: null,
     srcImg: null,
     patch: 0,
+    mu: 0.0,
+    sigma: 1.0,
     _: null,
     dispose: function(complete) {
         var self = this;
@@ -187,7 +276,34 @@ NNF.prototype = {
     clone: function() {
         return new NNF(this);
     },
-    normalize: function() {
+    scale: function(dst, src, scale) {
+        var self = this, A, B,
+            scaled = (new NNF(dst, src, self.patch)).init();
+        if (self._)
+        {
+            A = self.dst.points();
+            B = self.src.points();
+            self._.forEach(function(v, a) {
+                var n = stdMath.floor(scale), dx, dy, aa, bb, b = v[0], d = v[1],
+                    ax = stdMath.round(scale*A[a].x),
+                    ay = stdMath.round(scale*A[a].y),
+                    bx = stdMath.round(scale*B[b].x),
+                    by = stdMath.round(scale*B[b].y);
+                for (dy=0; dy<n; ++dy)
+                {
+                    for (dx=0; dx<n; ++dx)
+                    {
+                        aa = scaled.dst.indexOf(ax + dx, ay + dy);
+                        bb = scaled.src.indexOf(bx + dx, by + dy);
+                        if (-1 !== aa && -1 !== bb) scaled._[aa] = [bb, d];
+                    }
+                }
+            });
+        }
+        return scaled;
+    },
+    normalize: function(apply) {
+        if (!this._) return this;
         var self = this,
             _ = self._, n = _.length,
             a, b, d, sum, mu, sigma;
@@ -205,11 +321,21 @@ NNF.prototype = {
             sum += d * d;
         }
         sigma = stdMath.sqrt(sum / (n - 1));
-        for (a=0; a<n; ++a)
+        if (true === apply)
         {
-            b = _[a][0];
-            d = stdMath.abs((_[a][1] - mu) / sigma);
-            _[a] = [b, d];
+            for (a=0; a<n; ++a)
+            {
+                b = _[a][0];
+                d = stdMath.abs((_[a][1] - mu) / sigma);
+                _[a] = [b, d];
+            }
+            self.mu = 0.0;
+            self.sigma = 1.0;
+        }
+        else
+        {
+            self.mu = mu;
+            self.sigma = sigma;
         }
         return self;
     },
@@ -241,21 +367,21 @@ NNF.prototype = {
         {
             ya = ay+y; yb = by+y;
             if (0 > ya || 0 > yb || ya >= ah || yb >= bh) continue;
-            if (!AA.has(null, ya) || !BB.has(null, yb))
+            /*if (!AA.has(null, ya) || !BB.has(null, yb))
             {
                 excluded += patch;
                 continue;
-            }
+            }*/
             yaw = ya*aw; ybw = yb*bw;
             for (x=-p; x<=p; ++x)
             {
                 xa = ax+x; xb = bx+x;
                 if (0 > xa || 0 > xb || xa >= aw || xb >= bw) continue;
-                if (!AA.has(xa, null) || !BB.has(xb, null))
+                /*if (!AA.has(xa, null) || !BB.has(xb, null))
                 {
                     excluded += 1;
                     continue;
-                }
+                }*/
                 if (isRGBA)
                 {
                     i = (xa + yaw) << 2;
@@ -275,9 +401,9 @@ NNF.prototype = {
                 ++completed;
             }
         }
-        return /*20*excluded >= patch*patch ? false :*/ (completed ? (diff+excluded*65025/*255*255*/*(isRGBA?3:1))/(completed+excluded) : INF);
+        return /*20*excluded >= patch*patch ? false :*/ (completed ? (diff/*+excluded*65025/*255*255* /*(isRGBA?3:1)*/)/(completed/*+excluded*/) : INF);
     },
-    avg: function(bp, output) {
+    avg: function(bp, output, weight) {
         var self = this,
             B = self.src,
             dataB = self.srcImg,
@@ -289,7 +415,9 @@ NNF.prototype = {
             dx, dy, x, y, yw,
             r = 0.0, g = 0.0,
             b = 0.0, sum = 0.0,
-            i, index, p = patch >>> 1;
+            index, p = patch >>> 1;
+
+        weight = 1;
         if (4 === dataB.channels)
         {
             for (dy=-p; dy<=p; ++dy)
@@ -302,10 +430,10 @@ NNF.prototype = {
                     x = bx+dx;
                     if (0 > x || x >= width || !B.has(x, null)) continue;
                     index = (x + yw) << 2;
-                    r += imgB[index + 0];
-                    g += imgB[index + 1];
-                    b += imgB[index + 2];
-                    sum += 1;
+                    r += imgB[index + 0] * weight;
+                    g += imgB[index + 1] * weight;
+                    b += imgB[index + 2] * weight;
+                    sum += weight;
                 }
             }
             output[0] = clamp(stdMath.round(r / sum), 0, 255);
@@ -324,14 +452,85 @@ NNF.prototype = {
                     x = bx+dx;
                     if (0 > x || x >= width || !B.has(x, null)) continue;
                     index = (x + yw);
-                    r += imgB[index];
-                    sum += 1;
+                    r += imgB[index] * weight;
+                    sum += weight;
                 }
             }
             output[0] = clamp(stdMath.round(r / sum), 0, 255);
         }
         return output;
     },
+    /*conf: function(gamma) {
+        var self = this,
+            _ = self._,
+            A = self.dst.points(),
+            n = _.length, mapA = {},
+            width = self.dstImg.width,
+            height = self.dstImg.height,
+            i, j, ii, v, vals, ads,
+            value = 1.50, a = 1.0, b = 1.5,
+            result = new Array(n),
+            x, y, lx, ty, rx, by,
+            cs, cx, cy;
+        vals = [0,0,0,0];
+        ads = [b, a, b, a];
+        for (i=0; i<n; ++i) mapA[A[i].index] = i;
+        for (i=0; i<n; ++i)
+        {
+            x = A[i].x;
+            y = A[i].y;
+            lx = x - 1;
+            ty = y - 1;
+            rx = x + 1;
+            cs = [[lx, ty], [x, ty], [rx, ty], [lx, y]];
+            for (j=0; j<4; ++j)
+            {
+                cx = cs[j][0]; cy = cs[j][1];
+                if (0 <= cx && cx < width && 0 <= cy && cy < height)
+                {
+                    ii = mapA[(cx + cy * width)];
+                    v = null != ii ? (result[ii]||0) : 0;
+                    vals[j] = v + ads[j];
+                }
+                else
+                {
+                    vals[j] = a;
+                }
+            }
+            result[i] = stdMath.min.apply(stdMath, vals);
+        }
+        vals = [0,0,0,0,0];
+        for (i=n-1; i>=0; --i)
+        {
+            x = A[i].x;
+            y = A[i].y;
+            lx = x - 1;
+            by = y + 1;
+            rx = x + 1;
+            cs = [[lx, by], [x, by], [rx, by], [rx, y]];
+            for (j=0; j<4; ++j)
+            {
+                cx = cs[j][0]; cy = cs[j][1];
+                if (0 <= cx && cx < width && 0 <= cy && cy < height)
+                {
+                    ii = mapA[(cx + cy * width)];
+                    v = null != ii ? (result[ii]||0) : 0;
+                    vals[j] = v + ads[j];
+                }
+                else
+                {
+                    vals[j] = a;
+                }
+            }
+            vals[4] = result[i];
+            result[i] = stdMath.min.apply(stdMath, vals);
+        }
+        for (i=0; i<n; ++i)
+        {
+            result[i] = 0 == result[i] ? value : stdMath.pow(gamma, result[i]);
+        }
+        return result;
+    },*/
     init: function() {
         var self = this, _,
             AA = self.dst,
@@ -340,6 +539,7 @@ NNF.prototype = {
             B = BB.points(),
             n = A.length, a, b,
             res, tries;
+        if (!A.length || !B.length) {self._ = null; return self;}
         if (!self._) self._ = new Array(A.length);
         _ = self._;
         for (a=0; a<n; ++a)
@@ -431,6 +631,7 @@ NNF.prototype = {
         return self;
     },
     run: function(iters, alpha, radius) {
+        if (!this._) return this;
         var self = this, n = self._.length, rectB = self.srcImg.rect, i, a;
         radius = stdMath.min(radius, rectB.width/2, rectB.height/2);
         for (i=0; i<iters; ++i)
@@ -453,6 +654,7 @@ NNF.prototype = {
         return self;
     },
     apply: function(metrics) {
+        if (!this._) return this;
         var self = this,
             _ = self._,
             dataA = self.dstImg,
@@ -478,7 +680,7 @@ NNF.prototype = {
                     ap = A[i];
                     bp = B[_[i][0]];
                     ai = ap.index << 2;
-                    self.avg(bp, color);
+                    self.avg(bp, color, 1);
                     dr = imgA[ai + 0] - color[0];
                     dg = imgA[ai + 1] - color[1];
                     db = imgA[ai + 2] - color[2];
@@ -516,7 +718,7 @@ NNF.prototype = {
                     ap = A[i];
                     bp = B[_[i][0]];
                     ai = ap.index;
-                    self.avg(bp, color);
+                    self.avg(bp, color, 1);
                     dr = imgA[ai] - color[0];
                     nmse += (dr * dr) / 65025/*255*255*/;
                     imgA[ai] = color[0];
@@ -557,25 +759,7 @@ NNF.unserialize = function(dst, src, obj) {
     nnf._ = obj._;
     return nnf;
 };
-
-function patchmatch(dst, src, patch, iters, alpha, radius)
-{
-    var srcData = src.data(), srcRect = src.rect();
-    if (dst.data().data === srcData.data)
-    {
-        // use same image data, copy to avoid distortions
-        src = new FILTER.Util.Image.Selection(
-            copy(srcData.data), srcData.width, srcData.height, srcData.channels,
-            {
-            x:srcRect.from.x, y:srcRect.from.y, width:srcRect.width, height:srcRect.height,
-            points:src.points()
-            }
-        );
-    }
-    return (new patchmatch.NNF(dst, src, patch)).init().run(iters, alpha, radius);
-}
 patchmatch.NNF = NNF;
-FILTER.Util.Filter.patchmatch = patchmatch;
 
 function randInt(a, b)
 {